added comment-delete authorization

+++ app/Http/Controllers/CommentController.php
@@ -4,13 +4,19 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
+    public function __construct() {
+        $this->middleware('auth');
+    }
+
     public function create(Request $request) {
         $comment = new Comment;
         $comment->content = $request->input('content');
         $comment->article_id = $request->input('article_id');
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return back();
@@ -18,8 +24,12 @@ class CommentController extends Controller
 
     public function delete($id) {
         $comment = Comment::find($id);
-        $comment->delete();
-        
-        return back();
+
+        if (Gate::allows('comment-delete', $comment)) {
+            $comment->delete();
+            return back();
+        } else {
+            return back()->with('error', 'Unauthorize');
+        }
     }
 }
